retrieval_mech/requirements.txt
python=3.9
numpy=1.23.5
faiss-cpu
nltk
transformers
pytorch